Moving over value file and updating features
